fix: calculate user order for contributions at data load
This fixes performance and also the link to each user's view.

Test-bot: skip
